# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Code Compiler API",
    "version": "1.0.0",
    "description": "API for code compilation and execution"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
...
    "/users/register": {
      "post": {
        "summary": "Register new user",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Invalid email domain or user already exists" },
          "500": { "description": "User registration failed" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "User login",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["emailOrUsername", "password"],
                "properties": {
                  "emailOrUsername": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User logged in successfully" },
          "401": { "description": "Invalid credentials" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      },
      "patch": {
        "summary": "Update user profile",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User profile updated successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/users/change-password": {
      "put": {
        "summary": "Change user password",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["currentPassword", "newPassword"],
                "properties": {
                  "currentPassword": { "type": "string" },
                  "newPassword": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password changed successfully" },
          "400": { "description": "Current password is incorrect" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/users/payment-validation": {
      "post": {
        "summary": "Validate user's payment method",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["paymentMethodId"],
                "properties": {
                  "paymentMethodId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment method validated successfully" },
          "400": { "description": "Invalid payment method" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/users/client-credentials": {
      "get": {
        "summary": "Get user's client credentials",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Client credentials retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clientId": { "type": "string" },
                    "clientSecret": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Client credentials not found" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/users/credits": {
      "get": {
        "summary": "Get user credits information",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User credits retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "planCredits": {
                      "type": "string",
                      "description": "Number of credits from the user's plan, or 'Unlimited' for yearly plans"
                    },
                    "promotionalCredits": {
                      "type": "integer",
                      "description": "Number of promotional credits"
                    },
                    "purchasedCredits": {
                      "type": "integer",
                      "description": "Number of purchased credits"
                    },
                    "totalCredits": {
                      "type": "string",
                      "description": "Total number of available credits, or 'Unlimited' for yearly plans"
                    },
                    "subscriptionPlan": {
                      "type": "string",
                      "description": "Current subscription plan"
                    },
                    "subscriptionEnd": {
                      "type": "string",
                      "format": "date-time",
                      "description": "End date of the current subscription, null for free plans"
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/users/delete-account": {
      "delete": {
        "summary": "Delete user account",
        "tags": ["User"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Account deleted successfully" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" },
          "500": { "description": "Server error" }
        }
      }
    },
...
    "/subscription/status": {
    "get": {
      "summary": "Get subscription status",
      "tags": ["Subscription"],
      "security": [{ "bearerAuth": [] }],
      "responses": {
        "200": {
          "description": "Subscription status retrieved successfully",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionStatus"
              }
            }
          }
        },
        "401": { "description": "Unauthorized" },
        "500": { "description": "Server error" }
      }
    }
  },
  "/subscription/history": {
    "get": {
      "summary": "Get subscription history",
      "tags": ["Subscription"],
      "security": [{ "bearerAuth": [] }],
      "responses": {
        "200": {
          "description": "Subscription history retrieved successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SubscriptionHistory"
                }
              }
            }
          }
        },
        "401": { "description": "Unauthorized" },
        "500": { "description": "An error occurred while fetching subscription history" }
      }
    }
  },
  "/subscription/plans": {
    "get": {
      "summary": "Get all subscription plans with details",
      "tags": ["Subscription"],
      "responses": {
        "200": {
          "description": "All plans retrieved successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PlanDetails"
                }
              }
            }
          }
        },
        "500": { "description": "Server error" }
      }
    }
  },
  "/subscription/plans/{plan_id}/create-razorpay-order": {
    "post": {
      "summary": "Create a Razorpay order for subscription",
      "tags": ["Subscription"],
      "security": [{ "bearerAuth": [] }],
      "parameters": [
        {
          "name": "plan_id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "responses": {
        "200": { "description": "Razorpay order created successfully" },
        "400": { "description": "Invalid plan ID" },
        "401": { "description": "Unauthorized" },
        "500": { "description": "Server error" }
      }
    }
  },
  "/subscription/verify-razorpay-payment": {
    "post": {
      "summary": "Verify Razorpay payment for subscription",
      "tags": ["Subscription"],
      "security": [{ "bearerAuth": [] }],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["razorpay_order_id", "razorpay_payment_id", "razorpay_signature"],
              "properties": {
                "razorpay_order_id": { "type": "string" },
                "razorpay_payment_id": { "type": "string" },
                "razorpay_signature": { "type": "string" }
              }
            }
          }
        }
      },
      "responses": {
        "200": { "description": "Payment verified successfully" },
        "400": { "description": "Invalid signature" },
        "401": { "description": "Unauthorized" },
        "500": { "description": "Server error" }
      }
    }
  },
  "/subscription/plans/{plan_id}/upgrade": {
    "post": {
      "summary": "Upgrade to a plan",
      "tags": ["Subscription"],
      "security": [{ "bearerAuth": [] }],
      "parameters": [
        {
          "name": "plan_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": { "description": "Subscription upgraded successfully" },
        "400": { "description": "Invalid plan or insufficient balance" },
        "401": { "description": "Unauthorized" }
      }
    }
  },
  "/subscription/plans/{plan_id}/downgrade": {
      "post": {
        "summary": "Downgrade subscription",
        "tags": ["Subscription"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription downgraded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "subscription": { "$ref": "#/components/schemas/SubscriptionStatus" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid plan or downgrade not allowed" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        }
      }
    },
...
  "/subscription/cancel": {
    "post": {
      "summary": "Cancel current subscription",
      "tags": ["Subscription"],
      "security": [{ "bearerAuth": [] }],
      "responses": {
        "200": { "description": "Subscription cancelled successfully" },
        "404": { "description": "No active subscription found" },
        "401": { "description": "Unauthorized" },
        "500": { "description": "Server error" }
      }
    }
  },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PlanDetails": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "creditsPerDay": { "type": "number" },
          "price": { "type": "number" },
          "duration": { "type": "number" },
          "totalCredits": { "type": "number" },
          "paymentMethods": { 
            "type": "array",
            "items": { "type": "string" }
          },
          "termsAndConditions": { "type": "string" }
        }
      },
      "SubscriptionHistory": {
        "type": "object",
        "properties": {
          "plan": { "type": "string" },
          "status": { "type": "string", "enum": ["Active", "Inactive", "Cancelled", "Expired"] },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "creditsPerDay": { "type": "number" },
          "cancelledAt": { "type": "string", "format": "date-time" }
        }
      },
      "SubscriptionStatus": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "enum": ["Active", "Inactive", "Cancelled", "Expired"] },
          "plan": { "type": "string" },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "creditsPerDay": { "type": "number" },
          "remainingCredits": { "type": "number" },
          "cancelledAt": { "type": "string", "format": "date-time" },
          "message": { "type": "string" }
        }
      },
...
    }
  }
}